# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Coupang AI Detail Maker API
  version: 0.1.0
  description: |
    Coupang AI Detail Maker 백엔드 REST API 명세.

    - 기준 문서: PRD v0.3.9(8장 주요 API가 엔드포인트 원천), 도메인 정의서 v0.3.10, 구조 원칙 v0.1.10, ERD v0.1.10
    - 문서가 갱신되면 이 명세도 갱신한다.
    - JSON 필드는 camelCase(NM-10). 오류는 `{ "error": { "code", "message" } }` 하나로 통일(NM-11), code는 구조 원칙 4.2 표만 쓴다.
    - 판정 우선순위(FR-06, 구조 원칙 4.2): 401 → 404 → 409(PUBLISHED 대상) → 403(이메일 미인증) → 402(잔액 부족) → 409(version·진행 중 작업) → 429 → 503. 미인증과 잔액 0이 겹치면 403.
    - 레이트 리밋(NFR-04): 일반 API 사용자당 60 req/min(인증 뒤에 적용), 로그인 IP당 10 req/min, 토큰 갱신 IP당 30 req/min. 초과 시 429 `RATE_LIMITED`. 가입·로그아웃은 리밋 없음.
    - uuid 형식이 아닌 프로젝트 id는 404 `NOT_FOUND`.
    - 처리되지 않은 오류는 500 `INTERNAL`(message에 내부 정보 없음). 각 operation에는 따로 적지 않는다.
    - 퍼블리시 전 모든 응답에 원본·프리뷰 키, draftHtml, finalHtml, 공개 URL이 없다(PP-05, NFR-08).
    - 표기: 우선순위 M/S, 단계 P1(2일 핵심 슬라이스)·P2(MVP 완성). S는 MVP 이후(M4).
servers:
  - url: /
    description: 단일 도메인·동일 출처(DEC-02). Express가 frontend/dist와 /api를 함께 서빙한다.
...
security:
  - bearerAuth: []
paths:
  /api/auth/signup:
    post:
      tags: [auth]
      operationId: signup
      summary: 이메일/비밀번호 가입
      description: |
        M·P1, FR-01, FR-36, BR-01·BR-05. bcrypt 해시 저장, User와 Wallet(0)을 한 TX에서 생성, `email_verified=false`.
        성공 시 새 패밀리로 Refresh Token(쿠키)과 Access Token(body)을 발급한다.
        이미 가입된 이메일은 400 `VALIDATION_FAILED`(I-12 해소). 이메일은 trim·소문자로 정규화한다. 비밀번호는 8자 이상 72바이트 이하(N-10 해소). 레이트 리밋 없음.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '201':
          description: 가입 성공, 토큰 발급
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetRefreshCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: 이메일/비밀번호 로그인
      description: |
        M·P1, FR-01, FR-36, BR-01·BR-06. 성공 시 `refresh_tokens` 1행 + `rt` 쿠키 + body `{accessToken, expiresIn}`.
        틀린 비밀번호는 401 `INVALID_CREDENTIALS`, 토큰 발급 0건. IP당 10 req/min(NFR-04).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: 로그인 성공
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetRefreshCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Refresh Token 회전과 새 Access Token 발급
      description: |
        M·P2, FR-37, FR-38, FR-40, BR-06. `Origin`이 `FRONTEND_ORIGIN`이 아니면 403 `ORIGIN_FORBIDDEN`.
        `rt` 검증 → DB 행 유효 확인 → 한 TX에서 기존 행 폐기(`revoked_at`, `replaced_by`)·같은 패밀리 새 행 → 새 Access Token 반환.
        이미 회전·폐기된 `rt`가 오면 패밀리 전체를 폐기하고 401 `REFRESH_INVALID`(재사용 탐지). 패밀리 최초 발급 후 30일 초과도 401.
        앱 시작 시 1회, 401 `TOKEN_EXPIRED` 수신 시 1회 호출한다. IP당 30 req/min.
      security:
        - refreshCookie: []
      responses:
        '200':
          description: 회전 성공. 새 `rt`는 이전 값과 다르다
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SetRefreshCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: 로그아웃(Refresh Token 패밀리 폐기)
      description: |
        M·P2, FR-39, BR-06. 현재 패밀리를 폐기하고 `rt` 쿠키를 `Max-Age=0`으로 삭제한다. Origin 검사(403 `ORIGIN_FORBIDDEN`).
        남은 Access Token은 `exp`까지만 유효. `rt`가 없거나 무효여도 204(레이트 리밋 없음).
      security:
        - refreshCookie: []
      responses:
        '204':
          description: 로그아웃 완료
          headers:
            Set-Cookie:
              description: '`rt=; Max-Age=0; HttpOnly; Secure; SameSite=Strict; Path=/api/auth`'
              schema:
                type: string
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/projects:
    post:
      tags: [projects]
      operationId: createProject
      summary: 프로젝트 생성
      description: |
        M·P1(자격 검사는 P2 BE-03b), FR-10, FR-06, BR-38. DRAFT, version 1로 생성. body `form`은 선택(DEC-05).
        프로젝트가 아직 없으므로 라우트에서 `assertEligible`을 호출한다. 판정 순서: 401 → 403 → 402 → 429.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                form:
                  $ref: '#/components/schemas/ProjectForm'
      responses:
        '201':
          description: 생성됨(status=DRAFT, version=1)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    get:
      tags: [projects]
      operationId: listProjects
      summary: 내 프로젝트 목록
      description: |
        M·P2(BE-05b), FR-10, FR-20. 내 프로젝트만, PUBLISHED 포함. 자격 무관(잔액 0이어도 200).
        최신순(created_at DESC) 전체 배열, 페이지네이션 없음, 항목은 Project 스키마 그대로(N-3 일부 해소).
      responses:
        '200':
          description: 프로젝트 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/projects/{id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [projects]
      operationId: getProject
      summary: 프로젝트 상세
      description: M·P1(BE-05a), FR-10, FR-20, C-8. 남의 프로젝트는 404. PUBLISHED 포함, 자격 무관. 진행 중 새로고침 시 `activeJobType`으로 상태를 복원한다.
      responses:
        '200':
          description: 프로젝트
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/projects/{id}/form:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    put:
      tags: [projects]
      operationId: saveProjectForm
      summary: 폼 저장 `*`
      description: |
        M·P2(BE-05b), FR-10, FR-34, FR-35, BR-35·BR-48. DRAFT·ANALYZED에서만, 그 외 409 `INVALID_STATE`. 성공 시 version+1.
        필수값 검증(BR-35)은 하지 않는다(생성 요청 시점). 판정 순서: 404 → 409 PUBLISHED → 403 → 402 → 409 version·진행 중 작업 → 429.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [form, version]
              properties:
                form:
                  $ref: '#/components/schemas/ProjectForm'
                version:
                  $ref: '#/components/schemas/Version'
      responses:
        '200':
          description: 저장됨. 응답 version = 요청 version + 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/projects/{id}/assets:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [assets]
      operationId: uploadAsset
      summary: 이미지 업로드(요청당 1장)
      description: |
        M·P2(BE-06), FR-11, BR-36·BR-50, D-19, C-6. 요청당 1장, 장당 10MB 이하, jpg/png/webp, 프로젝트당 최대 10장.
        `sharp`로 실제 이미지인지 확인. 원본은 비공개 버킷, 폭 390px 이하 워터마크 사본은 별도 키. 응답은 asset id만(원본·프리뷰 키 없음).
        11번째·10MB 초과·허용 외 형식은 400, 저장 0건. 판정 순서: 404 → 409 PUBLISHED → 403 → 402 → 429.
        성공 201 `{id}`. 업로드는 version을 올리지 않고 진행 중 작업과 무관하다. 확인 필요: 이미지 삭제·교체 API 없음(I-9).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 이미지 1장(image/jpeg, image/png, image/webp, 10MB 이하). multipart 필드명은 `file`
      responses:
        '201':
          description: 업로드됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/projects/{id}/analyze:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [analysis]
      operationId: analyzeProject
      summary: 경쟁사 분석 `*`
      description: |
        M·P2(BE-13), FR-12, FR-29, FR-34, FR-35, BR-20~27·BR-47, D-15·D-27. DRAFT·ANALYZED에서만(그 외 409).
        URL은 `https://(www.|m.)coupang.com/vp/products/{숫자}`만, 쿼리 제거. 형식 오류 400은 시도로 세지 않는다.
        `analyze_count`(프로젝트당 3회, 실패 포함)를 선점하고 크롤링 실패·LLM 실패(502)여도 복원하지 않는다. LLM 미호출 거절(429 `DAILY_LLM_LIMIT`, 503)은 복원.
        ANALYZED에서 재분석 성공 시 선택을 비우고 DRAFT. 크롤링 원문은 저장·로그 금지(BR-22). 동기 응답.
        판정 순서: 404 → 409 PUBLISHED → 403 → 402 → 409 version·진행 중 작업 → 429 → 503.
        body는 `{url, version}`. 성공 응답은 분석 결과만이고 새 version(+1)은 포함하지 않으므로 클라이언트는 프로젝트를 재조회한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, version]
              properties:
                url:
                  type: string
                  description: 쿠팡 상품 URL(D-15)
                  example: https://www.coupang.com/vp/products/1234567890
                version:
                  $ref: '#/components/schemas/Version'
      responses:
        '200':
          description: 분석 완료, USP 후보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/projects/{id}/usps:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    put:
      tags: [analysis]
      operationId: saveSelectedUsps
      summary: USP 선택 저장 → ANALYZED `*`
      description: |
        M·P2(BE-13), FR-12, FR-34, FR-35, BR-24·BR-27. 후보 안 값만, 1개 이상 → `selectedUsps` 저장, ANALYZED, version+1.
        DRAFT·ANALYZED에서만(GENERATED 이후 409). 판정 순서: 404 → 409 PUBLISHED → 403 → 402 → 409 version·진행 중 작업 → 429.
        0개 저장, 중복 값, 후보 밖 값은 400 `VALIDATION_FAILED`(I-10 해소).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [selectedUsps, version]
              properties:
                selectedUsps:
                  $ref: '#/components/schemas/UspList'
                version:
                  $ref: '#/components/schemas/Version'
      responses:
        '200':
          description: 저장됨(status=ANALYZED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/projects/{id}/generate:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [generation]
      operationId: generateDetailPage
      summary: 최초 생성(분석 생략 포함) `*`
      description: |
        M·P1(텍스트 중심, BE-09a) / P2(선점·이미지 필수, BE-09b), FR-13, FR-14, FR-34, FR-35, BR-11·BR-24·BR-25·BR-30·BR-35·BR-39.
        DRAFT·ANALYZED에서만. 필수값(D-18: 제품명 1~100자, 카테고리, 소개글 10~1,000자, 이미지 1장 이상) 누락은 400, LLM 0회.
        GENERATE 선점 → TX 밖 MAIN 호출 → 정제 → version 재확인 후 draftHtml 저장·GENERATED·version+1. DRAFT에서는 `selectedUsps=[]`(분석 생략).
        동기 응답(최대 90초, C-7). 차감 없음. 실패·타임아웃은 502, 상태 불변.
        판정 순서: 404 → 409 PUBLISHED → 403 → 402 → 409 version·진행 중 작업 → 429 → 503.
      requestBody:
        $ref: '#/components/requestBodies/VersionOnly'
      responses:
        '200':
          description: 생성 완료, 워터마크 프리뷰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/projects/{id}/regenerate:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [generation]
      operationId: regenerateDetailPage
      summary: 재생성 `*`
      description: |
        M·P2(BE-11), FR-15, FR-34, FR-35, BR-11·BR-34·BR-39·BR-47. GENERATED·EDITING에서만.
        `regen_count < 3` + REGEN 선점(조건부 원자 UPDATE) → MAIN → draftHtml 교체(수동 편집 초기화), GENERATED, version+1.
        실패·429·503·version 변경 시 선점 복원. 4번째 요청은 429 `REGEN_LIMIT`, LLM 0회. 차감 없음.
        판정 순서: 404 → 409 PUBLISHED → 403 → 402 → 409 version·진행 중 작업 → 429 → 503.
      requestBody:
        $ref: '#/components/requestBodies/VersionOnly'
      responses:
        '200':
          description: 재생성 완료, 새 워터마크 프리뷰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/projects/{id}/preview:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [preview]
      operationId: getPreview
      summary: 워터마크 프리뷰 HTML과 편집용 blocks
      description: |
        M·P1(워터마크, BE-10a) / P2(data URI·blocks, BE-10b), FR-16, FR-17, BR-32·BR-50·BR-51·BR-53·BR-54, D-21, DEC-01·DEC-08.
        draftHtml의 이미지를 390px 프리뷰 사본 data URI로 바꾸고 최상단 사선 오버레이 + 블록별 반복 워터마크를 넣어 응답(저장 안 함).
        워터마크 해제 파라미터 없음. 자격 무관. 클라이언트는 `sandbox` iframe(`srcdoc`)으로만 렌더링.
        생성 전(draftHtml 없음) 요청은 409 `INVALID_STATE`. 응답에 현재 `version`을 포함한다.
      responses:
        '200':
          description: 프리뷰
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token(HS256, `typ=access`, 15분). 로그인·갱신 응답 body로 받아 메모리에만 보관. 만료 401 `TOKEN_EXPIRED`, 서명·`typ` 오류 401 `TOKEN_INVALID`.
    extensionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: (S) 확장 토큰(`typ=ext`, `prj`=projectId, 10분). `GET /final`, `POST /publish-report`에서만 허용.
    refreshCookie:
      type: apiKey
      in: cookie
      name: rt
      description: Refresh Token(HttpOnly·Secure·SameSite=Strict·Path=/api/auth, 14일, 회전). `/api/auth/refresh`·`/logout`은 Origin도 검사.
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      description: 프로젝트 ID. 남의 프로젝트는 404
      schema:
        type: string
        format: uuid
  headers:
    SetRefreshCookie:
      description: '`rt=<JWT>; HttpOnly; Secure; SameSite=Strict; Path=/api/auth`'
      schema:
        type: string
    RetryAfter:
      description: 재시도까지 대기 시간(초)
      schema:
        type: integer
  requestBodies:
    VersionOnly:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [version]
            properties:
              version:
                $ref: '#/components/schemas/Version'
  responses:
    BadRequest:
      description: 입력 검증 실패 `VALIDATION_FAILED`(편집 text에 태그, URL 패턴 외, 업로드 제한, 필수값 누락, 중복 이메일 가입, USP 0개 저장 등)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_FAILED
              message: 입력값을 확인하세요
    Unauthorized:
      description: 인증 실패. `TOKEN_EXPIRED`(만료, 1회 갱신 후 재시도) / `TOKEN_INVALID`(서명·`typ` 오류, 갱신 없이 로그인) / `INVALID_CREDENTIALS`(로그인 실패) / `REFRESH_INVALID`(refresh 무효·재사용·폐기)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: TOKEN_EXPIRED
              message: 토큰이 만료되었습니다
    PaymentRequired:
      description: 잔액 부족 `INSUFFICIENT_CREDIT`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: '`EMAIL_NOT_VERIFIED`(이메일 미인증, 잔액 0과 겹쳐도 403 우선) / `ORIGIN_FORBIDDEN`(Origin 불일치) / `NOT_PUBLISHED`(미퍼블리시 final 조회)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: '`NOT_FOUND`. 없는 리소스, 남의 프로젝트(존재 비노출), uuid 형식이 아닌 프로젝트 id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: '`INVALID_STATE`(허용되지 않는 상태: PUBLISHED 변경 요청, 생성 전 preview, DRAFT·ANALYZED 퍼블리시 등) / `VERSION_CONFLICT`(version 불일치) / `JOB_IN_PROGRESS`(진행 중 LLM 작업). 클라이언트는 프로젝트를 다시 조회'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VERSION_CONFLICT
              message: 다른 곳에서 먼저 변경되었습니다
    TooManyRequests:
      description: '한도 초과, LLM 미호출. `ANALYZE_LIMIT` / `REGEN_LIMIT` / `AI_EDIT_LIMIT` / `DAILY_LLM_LIMIT`(Asia/Seoul 자정 초기화, 초기화 시각 안내 필드는 확인 필요) / `RATE_LIMITED`(NFR-04)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGateway:
      description: '`UPSTREAM_FAILED`. 크롤링·LLM 호출 실패·타임아웃(90초). 상태 불변'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: '`LLM_BUSY`. LLM 세마포어·대기열 초과(NFR-03). LLM 미호출, 선점 복원. `Retry-After` 뒤 재시도'
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 구조 원칙 4.2 오류 코드
              enum:
                - VALIDATION_FAILED
                - TOKEN_EXPIRED
                - TOKEN_INVALID
                - INVALID_CREDENTIALS
                - REFRESH_INVALID
                - INSUFFICIENT_CREDIT
                - EMAIL_NOT_VERIFIED
                - ORIGIN_FORBIDDEN
                - NOT_PUBLISHED
                - NOT_FOUND
                - VERSION_CONFLICT
                - JOB_IN_PROGRESS
                - INVALID_STATE
                - ANALYZE_LIMIT
                - REGEN_LIMIT
                - AI_EDIT_LIMIT
                - DAILY_LLM_LIMIT
                - RATE_LIMITED
                - UPSTREAM_FAILED
                - LLM_BUSY
                - INTERNAL
            message:
              type: string
              description: 사용자 안내 문구. 내부 정보 노출 금지
    Version:
      type: integer
      minimum: 1
      description: 기준 프로젝트 version(낙관적 잠금, FR-34). 현재 값과 다르면 409
    Credentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          description: 8자 이상, UTF-8 72바이트 이하(bcrypt 한계, N-10 해소)
    TokenResponse:
      type: object
      required: [accessToken, expiresIn]
      properties:
        accessToken:
          type: string
          description: Access Token(JWT). 메모리에만 보관
        expiresIn:
          type: integer
          description: 만료까지 초(기본 900)
      example:
        accessToken: eyJhbGciOiJIUzI1NiJ9...
        expiresIn: 900
...
    ProjectForm:
      type: object
      description: 제품 정보(D-18). 저장 시에는 필수값을 검사하지 않고 생성 요청 시점에 검사한다(BR-35)
      properties:
        productName:
          type: string
          maxLength: 100
          description: 제품명(생성 시 필수, 1~100자)
        category:
          type: string
          maxLength: 50
          description: 카테고리(생성 시 필수, 1~50자). 서버는 자유 문자열로 받고 선택지 목록은 검증하지 않는다
        intro:
          type: string
          maxLength: 1000
          description: 소개글(생성 시 필수, 10~1,000자)
        toneGuide:
          type: string
          maxLength: 200
          description: 톤앤매너(선택, 200자 이하, 없으면 기본 톤)
    UspList:
      type: array
      description: USP 목록. 요소는 문자열(분석 후보는 최대 10개, 각 100자 이하)
      items:
        type: string
    Project:
      type: object
      description: '`toProject` 화이트리스트 매퍼 응답. draftHtml·finalHtml·원본/프리뷰 키는 포함하지 않는다(PP-05)'
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [DRAFT, ANALYZED, GENERATED, EDITING, PUBLISHED]
        version:
          type: integer
          minimum: 1
        form:
          $ref: '#/components/schemas/ProjectForm'
        selectedUsps:
          $ref: '#/components/schemas/UspList'
        analyzeCount:
          type: integer
          minimum: 0
          description: 분석 시도(실패 포함), 상한 3
        regenCount:
          type: integer
          minimum: 0
          description: 재생성, 상한 3
        aiEditCount:
          type: integer
          minimum: 0
          description: AI 수정 성공, 상한 3(S)
        aiEditFailCount:
          type: integer
          minimum: 0
          description: AI 수정 실패, 상한 6(S)
        activeJobType:
          type: string
          nullable: true
          enum: [ANALYZE, GENERATE, REGEN, AI_EDIT, null]
          description: 진행 중 LLM 작업. null이면 없음
        publishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      example:
        id: 3f1c2a9e-0b7d-4c1e-9a55-2d7e8f0a1b23
        status: EDITING
        version: 4
        form:
          productName: 무선 청소기
          category: 가전
          intro: 가볍고 강력한 흡입력의 무선 청소기입니다.
        selectedUsps: []
        analyzeCount: 0
        regenCount: 1
        aiEditCount: 0
        aiEditFailCount: 0
        activeJobType: null
        publishedAt: null
        createdAt: '2026-09-30T09:00:00Z'
    Asset:
      type: object
      description: 업로드 응답은 asset id만(원본·프리뷰 키, 버킷 호스트 없음)
      properties:
        id:
          type: string
          format: uuid
    AnalysisResult:
      type: object
      description: 분석 결과(AnalysisResult). 크롤링 원문 없음(BR-22)
      properties:
        sourceUrl:
          type: string
          description: 쿼리를 제거한 쿠팡 상품 URL
        uspCandidates:
          $ref: '#/components/schemas/UspList'
        analyzedAt:
          type: string
          format: date-time
    Preview:
      type: object
      description: 워터마크 프리뷰(저장 안 함). 이미지는 data URI만, 원본 키·공개 URL·finalHtml 없음
      properties:
        version:
          type: integer
          minimum: 1
          description: 현재 프로젝트 version(편집 응답의 "새 version", BE-12). GET preview 응답에도 포함
        html:
          type: string
          description: 워터마크가 합성된 프리뷰 HTML(sandbox iframe srcdoc용)
        blocks:
          type: array
          description: 편집용 블록·필드 목록(DEC-08). `data-block-id`·`data-edit-id`에서 추출
          items:
            type: object
            required: [blockId, fields]
            properties:
              blockId:
                type: string
              fields:
                type: array
                items:
                  type: object
                  required: [editId, text]
                  properties:
                    editId:
                      type: string
                      description: 블록 내 고유
                    text:
                      type: string
                      description: 현재 텍스트
      example:
        version: 4
        html: '<div style="width:780px">...PREVIEW ONLY...</div>'
        blocks:
          - blockId: b1
            fields:
              - editId: e1
                text: 가볍고 강력한 흡입력
...